+++ maxgui/src/components/common/CreateResource/ServerFormInput.vue
@@ -31,7 +31,7 @@
  * Use of this software is governed by the Business Source License included
  * in the LICENSE.TXT file and at www.mariadb.com/bsl11.
  *
- * Change Date: 2025-03-24
+ * Change Date: 2025-04-28
  *
  * On the date above, in accordance with the Business Source License, use
  * of this software will be governed by version 2 or later of the General